+++ src/dataset/mod.rs
@@ -1,4 +1,5 @@
-use crate::trs::{rules, ConstantFold, Math};
+use crate::structs::Ruleset;
+use crate::trs::{ConstantFold, Math};
 use colored::*;
 use csv::ReaderBuilder;
 use egg::{Pattern, RecExpr, Runner, Searcher};
@@ -34,7 +35,7 @@ pub fn generate_dataset_par(
 }
 
 /// For a given expression and goal, the function
-/// adds to the data vector the expression along 
+/// adds to the data vector the expression along
 /// with the minimal subset of rules it needs to be
 /// proved. Algorithm explained in the manuscript
 #[allow(dead_code)]
@@ -56,7 +57,8 @@ pub fn minimal_set_to_prove(
     let mut proved_once: bool = false;
     // let mut minimal_ruleset_len: usize;
     let mut rule;
-    let mut ruleset = rules(ruleset_id);
+    let binding = Ruleset::from(ruleset_id);
+    let mut ruleset = binding.rules().clone();
     let data_object;
     ruleset.shuffle(&mut rng);
     println!("Ruleset size == {}", ruleset.len());
@@ -124,7 +126,7 @@ pub fn minimal_set_to_prove(
 }
 
 /// Same as generate_dataset, but this time
-/// the expression are passed without goals 
+/// the expression are passed without goals
 /// The goal is assumed to be either 0 or 1.
 #[allow(dead_code)]
 pub fn generate_dataset_0_1_par(
@@ -159,7 +161,7 @@ pub fn generate_dataset_0_1_par(
 }
 
 /// Same as minimal_set_to_prove, but this time
-/// the expression are passed without goals 
+/// the expression are passed without goals
 /// The goal is assumed to be either 0 or 1.
 #[allow(dead_code)]
 pub fn minimal_set_to_prove_0_1(
@@ -177,7 +179,8 @@ pub fn minimal_set_to_prove_0_1(
     let end_0: Pattern<Math> = "0".parse().unwrap();
     let goals = [end_0.clone(), end_1.clone()];
     let mut proved_goal = "0/1".to_string();
-    let result = crate::trs::prove(-1, expression, -2, params, true, false);
+    let ruleset = Ruleset::from(ruleset_id);
+    let result = crate::trs::prove(-1, expression, &ruleset, params, true, false);
     if result.result {
         let mut runner;
         let mut id;
@@ -185,7 +188,7 @@ pub fn minimal_set_to_prove_0_1(
         let mut i: usize;
         let mut counter: usize;
         let mut rule;
-        let mut ruleset = rules(ruleset_id);
+        let mut ruleset = ruleset.rules().clone();
         let data_object;
         ruleset.shuffle(&mut rng);
         let mut ruleset_copy: Vec<egg::Rewrite<Math, ConstantFold>>;
@@ -268,13 +271,13 @@ pub fn minimal_set_to_prove_0_1(
 /// It starts by reading expressions from a CSV file
 /// that must have the format (id, expression) for each
 /// row. It reads batch_size expressions before passing
-/// them to generate_dataset_0_1_par, which generate the 
-/// dataset for this batch of expressions and store it in a 
+/// them to generate_dataset_0_1_par, which generate the
+/// dataset for this batch of expressions and store it in a
 /// json file. Until the vector of expressions end.
 /// The continue_from_expr parameter can be used in case of failure.
 /// For example if we have 1 million expressions, our batch_size
 /// is set to 1000, and the execution fails after generating the
-/// dataset of 200 batched, we should relaunch the execution starting 
+/// dataset of 200 batched, we should relaunch the execution starting
 /// from the 200*1000 = 200,000th expression.
 pub fn generation_execution(
     file_path: &OsString,
