+++ src/rules/mod.rs
@@ -1,3 +1,7 @@
+use egg::Rewrite;
+
+use crate::trs::{ConstantFold, Math};
+
 pub mod add;
 pub mod and;
 pub mod andor;
@@ -12,3 +16,39 @@ pub mod mul;
 pub mod not;
 pub mod or;
 pub mod sub;
+
+pub fn all_rules() -> Vec<Rewrite<Math, ConstantFold>> {
+    vec![
+        add::add(),
+        and::and(),
+        andor::andor(),
+        div::div(),
+        eq::eq(),
+        ineq::ineq(),
+        lt::lt(),
+        max::max(),
+        min::min(),
+        modulo::modulo(),
+        mul::mul(),
+        not::not(),
+        or::or(),
+        sub::sub(),
+    ]
+    .into_iter()
+    .flatten()
+    .collect()
+}
+
+pub fn arith_rules() -> Vec<Rewrite<Math, ConstantFold>> {
+    vec![
+        add::add(),
+        div::div(),
+        modulo::modulo(),
+        mul::mul(),
+        sub::sub(),
+    ]
+    .into_iter()
+    .flatten()
+    .collect()
+}
+
